Extract raw predicate compiler
